feat : [step3] 자동차 구현체 패키지 분리 및 I/O 책임 역할 분리
